refactor: CheaterRegistrationForm 분리
